implementation of be bd bc commands

+++ hyperdbg/hprdbghv/GlobalVariables.h
@@ -84,7 +84,7 @@ LIST_ENTRY g_BreakpointsListHead;
  * @brief Seed for setting id of breakpoints
  * 
  */
-UINT32 g_MaximumBreakpointId;
+UINT64 g_MaximumBreakpointId;
 
 /**
  * @brief Shows whether the debugger transparent mode
